pyproject.toml
[project]
name = "git-commit-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.66.0",
    "gitpython>=3.1.45",
    "google-genai>=1.33.0",
    "grpcio>=1.74.0",
    "grpcio-reflection>=1.74.0",
    "openai>=1.106.1",
    "protobuf==6.31.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
]
